requirements.txt
django
djangorestframework
django-cors-headers
requests
python-dotenv
PyYaml